tons of fixes
git-svn-id: svn://84.194.112.164/bengine@183 2ee6282c-0497-4b59-8b7c-bcac2105c061

+++ src/brainz/neuronsensorz.h
@@ -7,7 +7,8 @@ class NeuronSensorz
 		NeuronSensorz();
 		~NeuronSensorz();
 	
-		int output;
+		//int output;
+		float output;
 };
 
 #endif
